+++ examples/dgcnn_benchmark_segmentation.py
@@ -58,7 +58,7 @@ SCANNET_TRANSFORM = T.Compose(
 )
 SCANNET_INFERER_TRANSFORM = T.Compose(
     [
-        T.BBoxCenter(keys="block_bbox", dst_keys=DataKeys.BLOCK_CENTER),
+        T.BBoxCenter(keys=DataKeys.BLOCK_BBOX, dst_keys=DataKeys.BLOCK_CENTER),
         T.CopyItems(keys=DataKeys.POS, names=DataKeys.NORM_POS),
         T.DivideKey(keys=DataKeys.NORM_POS, div_keys=DataKeys.SCENE_MAX),
         T.SubtractKey(keys=DataKeys.POS, sub_keys=DataKeys.BLOCK_CENTER, axes=[0, 1]),